requirements.txt
flask
flask-login
flask-sqlalchemy
flask-mail
flask-wtf
python-dotenv
pymysql
apscheduler
nltk
wordcloud
matplotlib
Pillow
werkzeug
